refactor: PR피드백 반영
