+++ src/main/java/com/chatcode/controller/ArticleController.java
@@ -1,8 +1,9 @@
 package com.chatcode.controller;
 
-import com.chatcode.dto.ArticleRequestDTO.*;
-import com.chatcode.dto.BaseResponseDto;
+import com.chatcode.dto.article.ArticleRequestDto.*;
+import com.chatcode.dto.scrap.BaseResponseDto;
 import com.chatcode.service.ArticleService;
+import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.Valid;
 import java.util.List;
 import java.util.Optional;
@@ -21,24 +22,28 @@ public class ArticleController {
 
 
     @PostMapping("")
+    @Operation(summary = "Article 작성", description = "Article을 작성합니다.")
     public ResponseEntity<BaseResponseDto<Void>> createArticle(@Valid @RequestBody ArticleCreateRequestDTO params) {
         articleService.articleCreate(params);
         return ResponseEntity.ok(new BaseResponseDto<>(1, null, "성공적으로 게시글이 등록되었습니다."));
     }
 
     @PutMapping("/{articleId}")
+    @Operation(summary = "Article 수정", description = "Article을 수정합니다.")
     public ResponseEntity<BaseResponseDto<Void>> updateArticle(@PathVariable Long articleId, @RequestBody ArticleUpdateRequestDTO params) {
         articleService.articleUpdate(articleId, params);
         return ResponseEntity.ok(new BaseResponseDto<>(1, null, "업데이트 성공"));
     }
 
     @GetMapping("")
+    @Operation(summary = "Article 목록 조회", description = "Article 목록을 조회합니다.")
     public ResponseEntity<BaseResponseDto<List<String>>> getArticleList() {
         List<String> articleList = articleService.readArticleList();
         return ResponseEntity.ok(new BaseResponseDto<>(1, articleList, "게시글 목록 조회 성공"));
     }
 
     @GetMapping("/{articleId}")
+    @Operation(summary = "Article 내용 조회", description = "Article 내용을 조회합니다.")
     public ResponseEntity<BaseResponseDto<Object>> readArticleById(@PathVariable Long articleId) {
         Optional<String> contentText = articleService.readArticleById(articleId);
         if (contentText.isPresent()) {
@@ -48,6 +53,7 @@ public class ArticleController {
     }
 
     @DeleteMapping("/{articleId}")
+    @Operation(summary = "Article 삭제", description = "Article을 삭제합니다.")
     public ResponseEntity<BaseResponseDto<Void>> deleteArticleAndContent(@PathVariable Long articleId) {
         articleService.deleteArticle(articleId);
         return ResponseEntity.ok(new BaseResponseDto<>(1, null, "포스트 삭제 성공"));
