+++ src/main/java/com/chatcode/service/ArticleService.java
@@ -1,14 +1,10 @@
 package com.chatcode.service;
 
 import com.chatcode.domain.entity.Article;
-import com.chatcode.domain.entity.ArticleTag;
-import com.chatcode.domain.entity.Tag;
-import com.chatcode.dto.ArticleRequestDTO.ArticleCreateRequestDTO;
-import com.chatcode.dto.ArticleRequestDTO.ArticleUpdateRequestDTO;
+import com.chatcode.dto.article.ArticleRequestDto.ArticleCreateRequestDTO;
+import com.chatcode.dto.article.ArticleRequestDto.ArticleUpdateRequestDTO;
 import com.chatcode.exception.common.ContentNotFoundException;
-import com.chatcode.repository.article.TagRepository;
 import com.chatcode.repository.article.ArticleRepository;
-import com.chatcode.repository.article.ArticleTagRepository;
 import com.chatcode.repository.article.ArticleWriteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -21,9 +17,8 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class ArticleService {
     private final ArticleRepository articleRepository;
-    private final TagRepository tagRepository;
-    private final ArticleTagRepository articleTagRepository;
     private final ArticleWriteRepository articleWriteRepository;
+    private final ArticleTagService articleTagService;
 
     @Transactional
     public void articleCreate(ArticleCreateRequestDTO params) {
@@ -31,8 +26,8 @@ public class ArticleService {
 
         if (params.getTags() != null && !params.getTags().isEmpty()) {
             Article article = articleWriteRepository.findById(articleId)
-                    .orElseThrow(() -> new RuntimeException("Article not found"));
-            tagToArticle(article, params.getTags());
+                    .orElseThrow(() -> new ContentNotFoundException("Article not found"));
+            articleTagService.tagToArticle(article, params.getTags());
         }
     }
 
@@ -45,9 +40,9 @@ public class ArticleService {
         articleRepository.updateArticle(articleId, contentId, updateDTO);
 
         Article article = articleWriteRepository.findById(articleId)
-                .orElseThrow(() -> new RuntimeException("Article not found"));
+                .orElseThrow(() -> new ContentNotFoundException("Article not found"));
 
-        updateArticleTags(article, updateDTO.getTags());
+        articleTagService.updateArticleTags(article, updateDTO.getTags());
     }
 
     public List<String> readArticleList() {
@@ -61,65 +56,7 @@ public class ArticleService {
 
     @Transactional
     public void deleteArticle(Long articleId) {
-        List<ArticleTag> articleTags = articleTagRepository.findByArticleId(articleId);
-        articleTagRepository.deleteAll(articleTags);
-
-        articleTags.forEach(articleTag -> {
-            Tag tag = articleTag.getTag();
-            tag.setTagCount(tag.getTagCount() - 1);
-            tagRepository.save(tag);
-        });
-
+        articleTagService.removeTagsFromArticle(articleId);
         articleRepository.deleteArticle(articleId);
-    }
-
-    private void tagToArticle(Article article, List<String> tag) {
-        for (String tagName : tag) {
-            Tag tags = findOrCreateTag(tagName);
-            tagCount(tags);
-
-            ArticleTag articleTag = new ArticleTag();
-            articleTag.setArticle(article);
-            articleTag.setTag(tags);
-            articleTagRepository.save(articleTag);
-        }
-    }
-
-    private void updateArticleTags(Article article, List<String> tags) {
-        List<ArticleTag> existingTags = articleTagRepository.findByArticleId(article.getId());
-
-        for (ArticleTag articleTag : existingTags) {
-            if (!tags.contains(articleTag.getTag().getName())) {
-                articleTagRepository.delete(articleTag);
-                Tag tag = articleTag.getTag();
-                tag.setTagCount(tag.getTagCount() - 1);
-                tagRepository.save(tag);
-            }
-        }
-
-        for (String tagName : tags) {
-            Tag tag = findOrCreateTag(tagName);
-            if (existingTags.stream().noneMatch(at -> at.getTag().getName().equals(tagName))) {
-                ArticleTag articleTag = new ArticleTag();
-                articleTag.setArticle(article);
-                articleTag.setTag(tag);
-                articleTagRepository.save(articleTag);
-                tagCount(tag);
-            }
-        }
-
-    }
-
-    private Tag findOrCreateTag(String tagName) {
-        return tagRepository.findByName(tagName).orElseGet(() -> {
-            Tag newTag = new Tag();
-            newTag.setName(tagName);
-            newTag.setTagCount(0);
-            return tagRepository.save(newTag);
-        });
-    }
-    private void tagCount(Tag tag) {
-        tag.setTagCount(tag.getTagCount() + 1);
-        tagRepository.save(tag);
     }
 }
